requirements.txt
torch
tqdm
ollama
transformers
python-liquid
sentence-transformers
faiss-cpu
beir